Midterm [Car detection]/machine learning/pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = ""
authors = [
    {name = "TopSlayer",email = "118892404+Top-Slayer@users.noreply.github.com"}
]
requires-python = ">=3.9,<3.12"
dependencies = [
    "tensorflow (==2.12.0)",
    "keras (==2.12.0)",
    "numpy (==1.23.5)",
    "flask (>=3.1.0,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
